Updated Fellows 3D model
Did a large amount of work on the 3D model and hammered out most of the
arms and shell. This process included making models of all of the
components used in the robots, all of which have been saved as
individual files.

I do believe that I've made an error in the channel measurements as I
just can't get it to match up with Actobotics dimension drawings. When I
get my hands on the real-life physical part I'll be able to correct the
mistake if one exists.