Reel Counter Refinement
Modified the sensor bars to accept the QRE1113 Digital Line Sensor
B.O.B.. In the first experiment the sensor was placed to far from the
wheel do to operator error.

Also added some tabs to the sides and slots to the top and bottom. This
will help hold the pieces together during assembly.