Gave encoder wheel larger gaps
After learning how quadratures actually work it became obvious that the
encoder's slots are too small. This was corrected and all that is left
is to add the hub mount holes.

I also began work on the second version of the real counter, which is
far more simple than the first. Why build an encoder when you can just
use the passing components to block yo gates? (Because I really needed
an excuse to build an encoder and unlock it's tasty secrets!)