MediaLive Spectrum Experimints (wintergreen)
Got some code working that flashes the LEDs with the tempo of the
music.., sort of. I'd really like to play with this for another few
months before presenting but, barring the use of my magical stopwatch,
that just ain't gunna happen.

I suppose I could use the watch, but it seems like a rather frivolous
use of one of the few charges it has left.  Given the difficulty in
recharging the device, I'll probably save the uses it has left for the
next time I need to save the planet or sneak into a locker room.